Clean project & organize files
