+++ Lock/ConnectionBuildable.swift
@@ -1,0 +1,48 @@
+// ConnectionBuildable.swift
+//
+// Copyright (c) 2016 Auth0 (http://auth0.com)
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
+
+import Foundation
+
+
+/**
+ *  Allows to specify Lock connections
+ */
+public protocol ConnectionBuildable: Connections {
+
+    /**
+     Configure a database connection
+
+     - parameter name:             name of the database connection
+     - parameter requiresUsername: if the database connection requires username
+     - important: Only **ONE** database connection can be used so subsequent calls will override the previous value
+     */
+    mutating func database(name name: String, requiresUsername: Bool)
+
+    /**
+     Adds a new social connection
+
+     - parameter name:  name of the connection
+     - parameter style: style used for the button used to trigger authentication
+     - seeAlso: AuthStyle
+     */
+    mutating func social(name name: String, style: AuthStyle)
+}
