+++ Lock/DatabaseForgotPasswordPresenter.swift
