+++ Lock/DatabaseForgotPasswordView.swift
